added sdl-1.3 build support
git-svn-id: https://fheroes2.svn.sourceforge.net/svnroot/fheroes2/trunk@1577 4f904287-c015-0410-838d-95d3a2656228

+++ fheroes2/src/fheroes2/game/game.cpp
@@ -238,7 +238,7 @@ u16 Game::GetGameOverScores(void)
     return GetRating() * (200 - nk) / 100;
 }
 
-void Game::KeyboardGlobalFilter(u16 sym, u16 mod)
+void Game::KeyboardGlobalFilter(u32 sym, u16 mod)
 {
     Display & display = Display::Get();
